# docs/openapi.yaml
openapi: 3.1.0
info:
  title: klaxond API
  version: 0.14.26
  summary: Notification bridge backend API
  description: |
    klaxond receives monitoring webhooks, renders notification payloads, manages
    routing/inhibition policy, and exposes a zero-build admin UI.

    Browser sessions use the `klaxond_session` cookie plus `X-Klaxond-CSRF` on
    same-origin mutations. Scoped machine clients should use Bearer tokens.
    Sensitive browser mutations can also return `428 Precondition Required`
    with `X-Klaxond-Reauth: required`; call `/api/auth/reauth` or complete passkey
    reauthentication before retrying.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
...
  /api/auth/callback:
    get:
      tags: [Auth]
      operationId: oidcCallback
      summary: Complete the OIDC authorization-code callback.
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Session established and redirected to the original return URL.
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          description: OIDC discovery or token exchange failed.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: klaxond scoped token
      description: API keys and PATs created from `/api/auth/tokens`.
    sessionCookie:
      type: apiKey
      in: cookie
      name: klaxond_session
      description: Signed browser session cookie.
    csrfHeader:
      type: apiKey
      in: header
      name: X-Klaxond-CSRF
      description: Required on browser-originated mutations when using a session cookie.
    basicAuth:
      type: http
      scheme: basic
      description: Username/password auth for non-browser clients in basic or LDAP mode.
...
  responses:
...
    BadRequest:
      description: Invalid request body, unsupported severity or validation failure.
      content:
        text/plain:
          schema:
            type: string
...
security:
  - bearerAuth: []
  - sessionCookie: []